scraping pipeline/requirements.txt
selenium
undetected-chromedriver
python-dotenv
pandas
langdetect
schedule
pytz

